chore: change project folder architecture
