+++ app/dto/transaction_dto.php
@@ -1,0 +1,18 @@
+<?php
+class TransactionDTO
+{
+    public $value;
+    public $category;
+    public $date;
+    public $description;
+    public $type;
+
+    function __construct(array $array)
+    {
+        $this->value = $array['value'];
+        $this->category = $array['category'];
+        $this->date = new DateTime($array['date']);
+        $this->description = $array['description'];
+        $this->type = $array['type'];
+    }
+}
