+++ app/pages/404.php
@@ -1,0 +1,1 @@
+<h1>Essa página nao existe 🤕!</h1>
