+++ app/pages/home.php
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Home - MinhaDespesa</title>
+</head>
+
+<body>
+    Não há nada aqui ainda!
+</body>
+
+</html>
