+++ app/pages/register.php
@@ -1,0 +1,8 @@
+<h1>Registre-se</h1>
+<form action="">
+    <label for="email">E-mail</label>
+    <input type="text" name="email" id="">
+    <label for="password">Senha</label>
+    <input type="password" name="password" id="">
+    <input type="submit" value="Entrar">
+</form>
